modifications

+++ web/application/models/doctorModel.php
@@ -311,9 +311,10 @@ class doctorModel extends database
         return $u;
     }
     public function getArticles($userid){
-        if($this->Query("SELECT p.id, p.type, p.heading, p.description 
+        if($this->Query("SELECT p.id, p.type, p.heading, p.description,pa.url
                          from post p
-                         where p.type<? && p.type!=? && p.approval_status=? && p.author_id=?",[7,1,1,$userid])){
+                         inner join post_attachments pa on pa.post_id=p.id
+                         where p.type<? && p.type!=? && p.approval_status=? && p.author_id=?",[7,1,2,$userid])){
             $x=$this->fetchall();
             return $x;
         }
@@ -338,11 +339,14 @@ class doctorModel extends database
             $type = 6;
             break;
         }
-        $y=[$data['userid'],$type,$data['heading'],$data['content']]; 
-            if($this->Query("INSERT INTO post (author_id,type,heading,description) VALUES (?,?,?,?)",$y)){
-                 return true;
-            }
+        $y=[$data['userid'],$type,$data['heading'],$data['content'],$data['filename']]; 
+            
+        if($this->Query("INSERT INTO post (author_id,type,heading,description) VALUES (?,?,?,?);SET @last_id_in_table1 = LAST_INSERT_ID();INSERT INTO post_attachments (post_id,url) VALUES (@last_id_in_table1,?)",$y))
+        {
+             return true;
+        }
     }
+
     public function deleteArticle($id)
     {
         
@@ -361,13 +365,19 @@ class doctorModel extends database
                 }
             }     
         }  
-        if($this->Query("DELETE from post where id=?",[$id]))
-        { 
-            
-            if($this->rowCount()>0){
-                return true; 
-            }  
-        }  
+       
+                if($this->Query("DELETE from comments where post_id=?",[$id]))
+                { 
+                    if($this->Query("DELETE from post where id=?",[$id]))
+                    { 
+                        
+                        if($this->rowCount()>0){
+                            return true; 
+                        }  
+                    }  
+                }
+           
+       
         
         return false;
